+++ src/more_containers/types.py
@@ -1,1 +1,2 @@
 from ._uniquetuple import UniqueTuple  # noqa : F401
+from ._attydict import AttyDict  # noqa : F401